Ispravljanje GET metoda za transakcije, ogroman refactor
