+++ internal/bank/repository.go
@@ -1,16 +1,13 @@
 package bank
 
 import (
-	"context"
 	"database/sql"
 	"errors"
 	"fmt"
 	"log"
-	"math"
 	"time"
 
 	bankpb "github.com/RAF-SI-2025/Banka-3-Backend/gen/bank"
-	"github.com/RAF-SI-2025/Banka-3-Backend/gen/exchange"
 	"github.com/jackc/pgx/v5/pgconn"
 )
 
@@ -865,272 +862,6 @@ func (s *Server) getOwnerFromAccount(tx *sql.Tx, account string) (int64, error) 
 	}
 
 	return ownerID, nil
-}
-
-func (s *Server) ProcessPayment(from_account string, to_account string, amount int64, transaction_code int64, call_number string, reason string) (*Payment, *Currency, error) {
-
-	fromAcc, err := s.GetAccountByNumberRecord(from_account)
-	if err != nil {
-		return nil, nil, ErrAccountNotFound
-	}
-	toAcc, err := s.GetAccountByNumberRecord(to_account)
-	if err != nil {
-		return nil, nil, ErrAccountNotFound
-	}
-
-	var finalAmount = amount
-	var commission int64 = 0
-
-	// 1. Logika konverzije
-	if fromAcc.Currency != toAcc.Currency {
-		ctx := context.Background()
-		// EUR -> RSD
-		resp1, err := s.ExchangeService.ConvertMoney(ctx, &exchange.ConversionRequest{
-			FromCurrency: fromAcc.Currency,
-			ToCurrency:   "RSD",
-			Amount:       float64(amount),
-		})
-		if err != nil {
-			return nil, nil, fmt.Errorf("exchange error (hop 1): %v", err)
-		}
-		// RSD -> USD
-		resp2, err := s.ExchangeService.ConvertMoney(ctx, &exchange.ConversionRequest{
-			FromCurrency: "RSD",
-			ToCurrency:   toAcc.Currency,
-			Amount:       resp1.ConvertedAmount,
-		})
-		if err != nil {
-			return nil, nil, fmt.Errorf("exchange error (hop 2): %v", err)
-		}
-
-		commission = int64(math.Round(float64(amount) * commission_rate))
-		finalAmount = int64(math.Round(resp2.ConvertedAmount))
-	}
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return nil, nil, fmt.Errorf("start tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	// 2. Ažuriranje balansa
-	if fromAcc.Currency != toAcc.Currency {
-		// Razlicita valuta:
-
-		systemEmail := "system@banka3.rs"
-		// A. Skini platiocu (Source)
-		if _, err := s.DecreaseAccountBalance(tx, from_account, amount); err != nil {
-			return nil, nil, err
-		}
-
-		// B. Dodaj banci (Source)
-		_, err = tx.Exec(`UPDATE accounts SET balance = balance + $1 WHERE currency = $2 AND owner = (SELECT id FROM clients WHERE email = $3)`,
-			amount, fromAcc.Currency, systemEmail)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to credit bank source account: %w", err)
-		}
-
-		// C. Skini banci (Target)
-		_, err = tx.Exec(`UPDATE accounts SET balance = balance - $1 WHERE currency = $2 AND owner = (SELECT id FROM clients WHERE email = $3)`,
-			finalAmount, toAcc.Currency, systemEmail)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to debit bank target account: %w", err)
-		}
-
-		// D. Dodaj primaocu (Target)
-		if _, err := s.IncreaseAccountBalance(tx, to_account, finalAmount); err != nil {
-			return nil, nil, err
-		}
-
-	} else {
-		// Ista valuta: Direktno
-		if _, err := s.DecreaseAccountBalance(tx, from_account, amount); err != nil {
-			return nil, nil, err
-		}
-		if _, err := s.IncreaseAccountBalance(tx, to_account, amount); err != nil {
-			return nil, nil, err
-		}
-	}
-
-	// 3. Kreiraj zapis o plaćanju
-	payment, err := s.CreatePayment(tx, from_account, to_account, amount, finalAmount, commission, transaction_code, call_number, reason)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, nil, fmt.Errorf("commit: %w", err)
-	}
-
-	currency, err := s.getCurrencyByLabel(fromAcc.Currency)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get currency id: %v", err)
-	}
-
-	return payment, currency, nil
-}
-
-func (s *Server) CreateTransfer(fromAccount, toAccount string, amount int64) (*Transfer, error) {
-	if fromAccount == toAccount {
-		return nil, errors.New("cannot transfer to same account")
-	}
-
-	fromAcc, err := s.GetAccountByNumberRecord(fromAccount)
-	if err != nil {
-		return nil, err
-	}
-	toAcc, err := s.GetAccountByNumberRecord(toAccount)
-	if err != nil {
-		return nil, err
-	}
-
-	var finalAmount = amount
-	var exchangeRate = 1.0
-	var commission int64 = 0
-
-	// Multi-currency logic: Always route through RSD if currencies differ
-	if fromAcc.Currency != toAcc.Currency {
-		ctx := context.Background()
-
-		// Source -> RSD
-		resp1, err := s.ExchangeService.ConvertMoney(ctx, &exchange.ConversionRequest{
-			FromCurrency: fromAcc.Currency,
-			ToCurrency:   "RSD",
-			Amount:       float64(amount),
-		})
-		if err != nil {
-			return nil, fmt.Errorf("exchange error (source to RSD): %v", err)
-		}
-
-		// RSD -> Target
-		resp2, err := s.ExchangeService.ConvertMoney(ctx, &exchange.ConversionRequest{
-			FromCurrency: "RSD",
-			ToCurrency:   toAcc.Currency,
-			Amount:       resp1.ConvertedAmount,
-		})
-		if err != nil {
-			return nil, fmt.Errorf("exchange error (RSD to destination): %v", err)
-		}
-
-		commission = int64(float64(amount) * commission_rate)
-		finalAmount = int64(resp2.ConvertedAmount)
-		exchangeRate = resp2.ExchangeRate
-	}
-
-	currency, err := s.getCurrencyByLabel(fromAcc.Currency)
-	if err != nil {
-		return nil, err
-	}
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	if fromAcc.Balance < amount {
-		return nil, errors.New("insufficient funds")
-	}
-
-	row := tx.QueryRow(`
-    INSERT INTO transfers (
-        from_account, to_account, start_amount, end_amount,
-        start_currency_id, exchange_rate, commission, status
-    )
-    VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
-    RETURNING transaction_id, from_account, to_account,
-              start_amount, end_amount,
-              start_currency_id, exchange_rate,
-              commission, status, timestamp
-`, fromAccount, toAccount, amount, finalAmount, currency.Id, exchangeRate, commission, pending)
-
-	transfer, err := scanTransfer(row)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-	return transfer, nil
-}
-
-func (s *Server) ConfirmTransfer(transferID int64, verificationCode string) error {
-	if verificationCode == "" {
-		return errors.New("verification code required")
-	}
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return err
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	var t Transfer
-	err = tx.QueryRow(`
-		SELECT transaction_id, from_account, to_account, start_amount, end_amount, status
-		FROM transfers WHERE transaction_id = $1
-	`, transferID).Scan(&t.Transaction_id, &t.From_account, &t.To_account, &t.Start_amount, &t.End_amount, &t.Status)
-	if err != nil {
-		return err
-	}
-
-	if t.Status != pending {
-		return errors.New("transfer already processed")
-	}
-
-	// Fetch account currencies to determine if we need the bank intermediary
-	fromAcc, _ := s.GetAccountByNumberRecord(t.From_account)
-	toAcc, _ := s.GetAccountByNumberRecord(t.To_account)
-
-	if fromAcc.Currency != toAcc.Currency {
-		// Multi-currency => Involve Bank Accounts
-		// We use the system email from seed.sql to find the bank's accounts
-		systemEmail := "system@banka3.rs"
-
-		// 1. Debit Client (Source Currency)
-		res, _ := tx.Exec(`UPDATE accounts SET balance = balance - $1 WHERE number = $2 AND balance >= $1`, t.Start_amount, t.From_account)
-		if aff, _ := res.RowsAffected(); aff == 0 {
-			return errors.New("insufficient funds")
-		}
-
-		// 2. Credit Bank (Source Currency)
-		_, err = tx.Exec(`UPDATE accounts SET balance = balance + $1 WHERE currency = $2 AND owner = (SELECT id FROM clients WHERE email = $3)`,
-			t.Start_amount, fromAcc.Currency, systemEmail)
-		if err != nil {
-			return err
-		}
-
-		// 3. Debit Bank (Target Currency)
-		_, err = tx.Exec(`UPDATE accounts SET balance = balance - $1 WHERE currency = $2 AND owner = (SELECT id FROM clients WHERE email = $3)`,
-			t.End_amount, toAcc.Currency, systemEmail)
-		if err != nil {
-			return err
-		}
-
-		// 4. Credit Client (Target Currency)
-		_, err = tx.Exec(`UPDATE accounts SET balance = balance + $1 WHERE number = $2`, t.End_amount, t.To_account)
-		if err != nil {
-			return err
-		}
-
-	} else {
-		// Same currency: Standard direct transfer
-		res, _ := tx.Exec(`UPDATE accounts SET balance = balance - $1 WHERE number = $2 AND balance >= $1`, t.Start_amount, t.From_account)
-		if aff, _ := res.RowsAffected(); aff == 0 {
-			return errors.New("insufficient funds")
-		}
-		_, err = tx.Exec(`UPDATE accounts SET balance = balance + $1 WHERE number = $2`, t.Start_amount, t.To_account)
-		if err != nil {
-			return err
-		}
-	}
-
-	_, err = tx.Exec(`UPDATE transfers SET status = $1 WHERE transaction_id = $2`, realized, t.Transaction_id)
-	if err != nil {
-		return err
-	}
-	return tx.Commit()
 }
 
 func (s *Server) GetTransferHistory(clientEmail string, page, pageSize int32) (*bankpb.TransferHistoryResponse, error) {
