+++ proto/bank/bank.proto
@@ -199,28 +199,30 @@ message UpdateCompanyRequest {
 message UpdateCompanyResponse {
   Company company = 1;
 }
+
 message Transaction {
   int64 id = 1;
-  string type = 2;
+  string type = 2; // da li je transfer ili placanje
   string from_account = 3;
   string to_account = 4;
-  double start_amount = 5;
-  double end_amount = 6;
-  double commission = 7;
-  string status = 8;
-  int64 timestamp = 9;
-
-  int64 recipient_id = 10;
-  string transaction_code = 11;
-  string call_number = 12;
-  string reason = 13;
-
-  int64 start_currency_id = 14;
-  double exchange_rate = 15;
+  double initial_amount = 5;
+  double final_amount = 6;
+  double fee = 7;
+  string currency = 8;
+  string payment_code = 9;
+  string reference_number = 10;
+  string purpose = 11;
+  string status = 12;
+  string timestamp = 13;
+  // placanje polja:
+  int64 recipient_id = 14;
+  // transfer polja:
+  int64 start_currency_id = 15;
+  double exchange_rate = 16;
 }
 
 message GetTransactionsRequest {
-  int64 client_id = 1;
+  string account_number = 1;
 
   string date_from = 2;
   string date_to = 3;
@@ -246,9 +248,8 @@ message GetTransactionsResponse {
 }
 
 message GetTransactionByIdRequest {
-  int64 client_id = 1;
-  int64 id = 2;
-  string type = 3;
+  int64 id = 1;
+  string type = 2;
 }
 
 message GetTransactionByIdResponse {
@@ -256,9 +257,8 @@ message GetTransactionByIdResponse {
 }
 
 message GenerateTransactionPdfRequest {
-  int64 client_id = 1;
-  int64 id = 2;
-  string type = 3;
+  int64 id = 1;
+  string type = 2;
 }
 
 message GenerateTransactionPdfResponse {
